+++ neuron/src/app/Neuron/CLI/App.hs
@@ -30,9 +30,11 @@ import qualified Neuron.LSP as LSP
 import qualified Neuron.Version as Version
 import Options.Applicative
 import Relude
+import System.Console.ANSI (hSupportsANSI)
 import System.Directory (getCurrentDirectory)
+import System.IO (hIsTerminalDevice)
 
-run :: (Bool -> App ()) -> IO ()
+run :: (GenCommand -> App ()) -> IO ()
 run act = do
   defaultNotesDir <- getCurrentDirectory
   cliParser <- commandParser defaultNotesDir <$> now
@@ -41,7 +43,8 @@ run act = do
       info
         (versionOption <*> cliParser <**> helper)
         (fullDesc <> progDesc "Neuron, future-proof Zettelkasten app <https://neuron.zettel.page/>")
-  let logAction = Logging.mkLogAction
+  useColors <- liftA2 (&&) (hIsTerminalDevice stderr) (hSupportsANSI stderr)
+  let logAction = Logging.mkLogAction useColors
   runApp (Env app logAction) $ runAppCommand act
   where
     versionOption =
@@ -52,21 +55,21 @@ run act = do
       tz <- getCurrentTimeZone
       utcToLocalTime tz <$> liftIO getCurrentTime
 
-runAppCommand :: (Bool -> App ()) -> App ()
+runAppCommand :: (GenCommand -> App ()) -> App ()
 runAppCommand genAct = do
   getCommand >>= \case
     LSP -> do
       LSP.lspServer
-    Gen GenCommand {..} -> do
-      case serve of
-        Just (host, port) -> do
+    Gen (mserve, gen) -> do
+      case mserve of
+        Just (ServeCommand host port) -> do
           outDir <- getOutputDir
           appEnv <- getAppEnv
           liftIO $
-            race_ (runApp appEnv $ genAct watch) $ do
+            race_ (runApp appEnv $ genAct gen) $ do
               runApp appEnv $ Backend.serve host port outDir
         Nothing ->
-          genAct watch
+          genAct gen
     New newCommand ->
       newZettelFile newCommand
     Open openCommand ->